requirements_Version2.txt
fastapi>=0.90
uvicorn[standard]>=0.20
pdfplumber>=0.6
sentence-transformers>=2.2
faiss-cpu>=1.7.4
joblib>=1.1
pandas>=1.3
numpy>=1.21
transformers>=4.30
torch>=1.12
openai>=0.27
python-multipart>=0.0.6
pytest>=7.0